feat: Add User and Post interfaces
This commit adds the `User` and `Post` interfaces to the `src/app/types` directory. The `User` interface includes properties for `name`, `image`, and `posts`, while the `Post` interface includes properties for `id`, `date`, `tag`, `title`, `body`, `image`, and `userName`. These interfaces will be used to define the shape of user and post objects throughout the application.